move deformatting utils to own file
